Change the type of 'Quantity' and 'price' from Float to Decimal

+++ superx/models/__init__.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Integer, Column, Text, Boolean, Float, BigInteger
+from sqlalchemy import Integer, Column, Text, Boolean, BigInteger, DECIMAL
 from flask_login import LoginManager, UserMixin
 from app import db
 
@@ -40,7 +40,7 @@ class Product(db.Model):
 
     id = Column(BigInteger, primary_key=True)
     name = Column(Text)
-    quantity = Column(Float)
+    quantity = Column(DECIMAL)
     is_weighted = Column(Boolean)
     unit_of_measure = Column(Text)
 
@@ -51,7 +51,7 @@ class BranchPrice(db.Model):
     branch_price_id = Column(Integer, primary_key=True, autoincrement=True)
     item_code = Column(db.ForeignKey('product.id'))
     branch_id = Column(db.ForeignKey('branch.id'))
-    price = Column(Float)
+    price = Column(DECIMAL)
     update_date = Column(Text)
